Restore assets before unit-tests are run (#32810)
## Issue
Fixes https://github.com/Azure/azure-sdk-for-js/issues/32805
Related PR: https://github.com/Azure/azure-sdk-for-js/pull/32806

The original issue was that "restoring recordings when the first test is
triggered" takes time, leading to flakily failing unit-test pipelines.
There is a need to have the recordings pre-loaded before the tests are
run in the CI pipeline.

#32806 - Earlier attempts of pre-restoring recordings in CI for playback
tests with retries and increased timeouts to avoid race conditions did
not fully solve the problem. This PR is a follow up of that.

## Description
Employs a new approach to restore assets sequentially using the
test-proxy tool, which is already downloaded by the CI pipeline.
The assets restoration process is applied only to the unit-test pipeline
in playback mode.

Updates to the `rush-runner` tool:

1. Enhanced `rushRunAllWithDirection` to handle test-proxy restore for
packages in CI pipeline.
2. Introduced new helper function: `runTestProxyRestore`.
3. Added a new helper function `spawnNodeWithOutput` to spawn NodeJS
programs and return the output.
4. Introduced `--ci` flag to help with debugging in local; calls `--ci`
flag in the yml scripts for the unit-test commands.

## Example Output
[Pipeline:
https://dev.azure.com/azure-sdk/public/_build/results?buildId=4523710&view=logs&j=45d25fdd-5540-5f10-8daf-622e80647be7&t=a7f7761c-0312-5a29-1d0c-9736ead6db1f&l=60](https://dev.azure.com/azure-sdk/public/_build/results?buildId=4529992&view=results)