[EngSys] add VS Code launch.json configurations for current file
It appears that with `rushx` we can run NPM scripts under sub directories. This
PR adds debug launch configurations that start debugging current active test
file in the editor.

Need to have `*-customized:current-file` configurations because some packages
pass additional arguments with `--` already so it would be wrong to `"--"` in
the `"args"` again. Users will need to pick one to use based on their test
scripts.