Filter ExcludePaths out of the targetedFiles (#33047)
This is the same change that was made in Python, is in my testing code
for Java and will be made for Rust.

The targetedFiles list needs to remove anything that's in the
ExcludePaths otherwise, it will end up trying to process them if they're
in the ChangedFiles or DeletedFiles. The scenario is this, something
changes in the root of path being excluded, the code on line 74 would
end up adding additional packages for validation for something that
should be being excluded.