[eslint] move monitor-opentelemetry prereqs building to its package
monitor-opentelemetry enables typed linting and needs to have its dependencies
built in order to have proper typing.  Currently we include the build steps for
every package in the analyze step, which is unnecessary for other packages.

This PR moves the prereqs build step to monitor-opentelemetry's package.json,
before running the `eslint` command.