[dev-tool] Remove env-paths dependency (#33130)
### Packages impacted by this PR

- @azure/dev-tool

### Issues associated with this PR

- #32939

### Describe the problem that is addressed by this PR

Removes the `env-paths` dependency from the @azure/dev-tool project and
only uses the logic we needed to get the cache location. Removed
dependency due to ESM only nature and the little benefit we got from it.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_


### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)