[Identity] Refactor MSAL Browser Flow  (#32060)
### Packages impacted by this PR
@azure/identity #31941

### Describe the problem that is addressed by this PR
This PR continues the MSAL refactoring for the browser flows based on
the work done by Maor in #28493. Some of the main changes for the PR
include:
- Provide a wrapper for the `MsalBrowserClient`
- Remove the dependency hierarchy between the interface and abstract
class

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_


### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)