[eventgrid] release v5.11.0 Feb 2025 (#33074)
### Packages impacted by this PR
`@azure/eventgrid`

### Issues associated with this PR

https://github.com/Azure/azure-rest-api-specs/pull/32013
https://github.com/Azure/azure-rest-api-specs/pull/32548

### Describe the problem that is addressed by this PR

This PR includes the following changes

- properties added to `AcsMessageMediaContent`
  - `animated`
- properties added to `AcsMessageReceivedEventData`
  -  `messageId`
  - `messageType`
  - `reaction`
  
### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

This is straight code regeneration and minor fix to adopt code-gen
changes.

### Are there test cases added in this PR? _(If not, why?)_

No. This item is standard and we need not add test cases for every new
events. The existing cases would be sufficient.

### Provide a list of related PRs _(if any)_

https://github.com/Azure/azure-sdk-for-js/pull/32441

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

`npm run generate:client`

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)