Add trailing slash exclude paths (#33156)
This PR is in preparation for an upcoming change that'll allow us to
exclude specific files. Right now, we were always adding a trailing "/"
if one wasn't there and this was done to ensure that things didn't
overmatch. For example, if you had sdk/foo and sdk/foobar with an
exclude of sdk/foo, without the slash, it would overmatch sdk/foo and
sdk/foobar. We were adding the trailing the slash to ensure that this
didn't happen. Because always have a need to exclude specific files,
adding the trailing slash won't allow this case to work.

This is the first change and it's a safe. Adding the trailing slash
allows the removal of where we were doing this in Language-Settings.ps1.
Note: There's still a place in eng/common Package-Properties.ps1 where
this also has to be done before specific file excludes can be added.