[core-client-rest] Implement environment checks for streaming methods (#33138)
### Packages impacted by this PR

- `@azure-rest/core-client`
- `@typespec/ts-http-runtime`

### Issues associated with this PR

- #32954

### Describe the problem that is addressed by this PR

The result of calling `asNodeStream` in the browser and of calling
`asBrowserStream` in Node is has the incorrect type. Since these methods
were only ever intended to be called in their corresponding
environments, this PR updates them to throw an error if they are called
in the wrong environment. Also updated docs to clarify.