[translation] Migrate translation projects to use snippets extraction (#33220)
### Packages impacted by this PR

- @azure-rest/ai-translation-document
- @azure-rest/ai-translation-text

### Issues associated with this PR

- https://github.com/Azure/azure-sdk-for-js/issues/32416

### Describe the problem that is addressed by this PR

Updates all projects under `translation` to use snippets extraction.
Note this disables non-functioning tests that need to be recorded under
the `documentTranslationTest.spec.ts` and
`translationFilterTest.spec.ts`. Noted in the issue #31948 that the
tests need to be re-recorded.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_


### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)