[dev-tool] Work around two minor issues (#33250)
### Packages impacted by this PR

`dev-tool`

### Describe the problem that is addressed by this PR

One issue is VS Code using up a ton of memory and the TS language
service crashing when viewing a `.js` file. This was fixed by adding a
`jsconfig.json` file. I figured this out from this comment:
https://github.com/microsoft/vscode/issues/235547#issuecomment-2547765610

The other issue is I wanted to be able to run dev-tool from outside of a
project folder when using admin commands (like listing all projects,
checking migration progress, etc) but `register.js` always tries to load
a local `package.json` and throws if you are in a folder without one so
I put a simple check to see if the file exists.