[EngSys] pin dev dependency `typescript` version to ~5.7.3 (#33325)
as 5.8 introduces some changes that cause generated .d.ts files incompatible
with previous versions of TypeScript compiler:

https://github.com/microsoft/TypeScript/issues/61360#issuecomment-2703357823

This PR temporarily pins `typescript` version to ~5.7.3 until we can upgrade to v5.8


### Issues associated with this PR

tracking upgrading to 5.8:
https://github.com/Azure/azure-sdk-for-js/issues/33324

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

We could upgrade to 5.8, but the incompatible output would cause a lot
of warnings from api-extractor because it still bundles 5.7.3.