[copilot] Improve GH copilot for JS (#33291)
This PR adds some instructions for GH copilot to improve its responses
and understanding of the Azure SDK for JS. This has worked well for me so
far, but definitely open to suggestions on what to tweak!

In this short demo, I purposely gave pretty vague instructions: how do I
add support for batch adding app configuration values - a new service
feature that will be added

And added entire codebase as context, again to be vague. 


https://github.com/user-attachments/assets/86e3ba0c-6d5b-4787-9db8-027e0643aa7a

---------

Co-authored-by: Jeremy Meng <jeremy.ymeng@gmail.com>