[core-rest-pipeline] remove `[custom]` from RestError members
TypeScript version v5.8 and later generate a type declaration for `RestError` that is incompatible with
previous versions of TypeScript, which means our d.ts generated by TypeScript v5.8+ will cause compilation
error when being consumed by earlier versions of TypeScript.

The `[util.custom]()` member of `RestError` is used to override the behavior of `util.inspec` on `RestError`
instances and not meant to be used directly. This PR removes the method from the class members and instead
assign to the `util.custom` property in the constructor.