[service bus] Disable local auth when not in TME (#33412)
`supportsSafeSecretStandard` is a parameter that tells whether the
subscription the resource is being created in disallows API-key-like
authentication. This PR updates the test-resources.json file to use this
parameter to disable such authentication methods if they're disallowed
by the target subscription. This unblocks creating resources in the
playground subscription.

Live run:
https://dev.azure.com/azure-sdk/internal/_build/results?buildId=4659763&view=results