[dev-tool] fix `migrate-package` to not format empty test directory
Part of SDK automation in specs repo PR check is to generate JS SDK packages for
spec changes and build them. The dev-tool `migrate-package` is executed on the
generated package before building it. There are case where the generated package
doesn't include a `test` directory, which causes "npm run format" to fail.

This PR checks whether the `test` directory exists or is empty. If either is
true test file patterns are removed from the `"format"` NPM script of the
package.