[@azure/cosmos] Change EncryptionQueryBuilder signature (#33426)
### Packages impacted by this PR
@azure/cosmos
### Issues associated with this PR


### Describe the problem that is addressed by this PR
This PR updates the signature of EncryptionQueryBuilder. A new interface
`CosmosEncryptedNumber` is created to accept numbers and a common
`addParameter` is added to take inputs.

This PR also adds a basic check to EncryptionKeyTimeToLiveInSeconds to
be >=60s as that is minimum interval at which cache gets cleared.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_
Yes

### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)

---------

Co-authored-by: Aman <aman@Amans-MacBook-Pro.local>