[service bus] Update tests to use secrets safely (#33431)
Updates the SB tests as follows:
- Currently the connection string is expected to be loaded from a .env
saved on disk. This PR changes this by accessing the connection strings
safely using ARM requests authenticated using Entra ID
- Updates connection-string-based tests to be skipped if the connection
string is not available

Additionally, it deletes a sample that was using a SAS key which is a
risky auth method that we should no longer advertise. It also refreshes
the samples.

Live tests run:
https://dev.azure.com/azure-sdk/internal/_build/results?buildId=4666221&view=results