[Search] safe secrets (#33437)
`supportsSafeSecretStandard` is a parameter that tells whether the
subscription the resource is being created in disallows API-key-like
authentication. This PR updates the `test-resources.bicep` file to use
this parameter to disable such authentication methods if they're
disallowed by the target subscription. This unblocks creating resources
in the playground subscription.

Live tests run:
https://dev.azure.com/azure-sdk/internal/_build/results?buildId=4666974&view=results

Failures in CI and nightly are unrelated to the PR.