[monitor] Migrate monitor-opentelmetry to ESM and vitest (#32819)
### Packages impacted by this PR

@azure/monitor-opentelemetry

### Issues associated with this PR

#31338 

### Describe the problem that is addressed by this PR

Migrates `@azure/monitor-opentelemetry` to ESM and vitest 


### Callouts

I fixed the majority of tests, but a few had to be skipped for this PR
due to various issues:

- the tests are failing in main, but the result is not awaited, so it
looks like they're passing. vitest is much better about catching
unhandled rejections and marking them as errors
- a few tests are passing in isolation, but failing when run in a group
(like `traceHandler`) which I believe is due to test pollution. It's
been a challenge fixing all the test pollution due to the many globals
used here so I will need to defer that work after this is shipped