[test] use custom vitest reporter (#33530)
### Packages impacted by this PR

- all packages using vitest

### Describe the problem that is addressed by this PR

The vitest `verbose` reporter outputs to console a JSON-serialized
version of any error thrown during a test run. In the case a `RestError`
is thrown, this results in the request and response objects being
output, including any secrets that may be present. This may lead to
secrets being leaked in CI logs. [I
thought](https://github.com/Azure/azure-sdk-for-js/pull/29737#issuecomment-2179047796)
that vitest turned off this behavior by default, but it turns out that
they left it on for the verbose reporter specifically for whatever
reason, so [when we changed
over](https://github.com/Azure/azure-sdk-for-js/pull/30581) to it from
the basic reporter back in July we inadvertently turned it back on.

So that we can still get the benefits of the verbose reporter while not
having vitest leak secrets, this PR updates our config to use a custom
reporter which updates a field on the verbose reporter to disable
outputting the serialized error.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

We could use another reporter, e.g. the default reporter, which doesn't
log the serialized error. But we updated to use the verbose reporter
because we find the output useful so why change back if we don't have
to? If this custom reporter proves to be annoying then maybe we can do
that instead.