[monitor] Re-enable traceHandler tests (#33549)
### Packages impacted by this PR

@azure/monitor-opentelemetry

### Issues associated with this PR

N/A - followup for #32819

### Describe the problem that is addressed by this PR

The move to ESM and Vitest uncovered some test-pollution that needs to
be addressed. To get the PR merged, I temporarily skipped those tests. 

This PR re-enables those tests to the best we can do given the
constraints in place and OTel's usage of globals.

A few changes were made:

1. OTel require hook will fire only once so instead of passing the
config to a new instrumentation, we can spy on the initialized instrumentation and
update its config via a spy. That allows it to always get a fresh config.
2. For ensuring disable works, we don't need to check on any spies and
it's sufficient to ensure that no http instrumentation was created.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

IMO this entire test suite can be simplified to only test the business
logic of the handler. When we initialize a handler, did it pass the correct
configuration to OTel? That will continue to test at the unit level with some
integration level coverage loss.