[core-lro] Relaxed status transformation (#33625)
### Packages impacted by this PR
@azure/core-lro

### Issues associated with this PR
N/A

### Describe the problem that is addressed by this PR
Polling responses may contain a status field, this PR logs it before it
transforms it to the set of known statuses. It also relaxes the
termination checking to be able to handle statuses that are not exactly
"succeeded" or "failed" but still contain some variation of these words.
This enables polling for many non-conforming LRO APIs without having to
create dedicated custom pollers for them.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

- Add an option that specifies the set of terminal statuses to check
for, this set has to be specified either by the user or by code gen.
This doesn't work in cases where this set is not part of the spec or the
user is not familiar with it
- Do nothing and let the user/code gen build custom pollers but that
feels a lot more expensive than it should be

### Are there test cases added in this PR? _(If not, why?)_
Yes

### Provide a list of related PRs _(if any)_
N/A

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [x] Added a changelog (if necessary)