[core-client-rest] Graceful error processing (#33623)
### Packages impacted by this PR
@azure-rest/core-client

### Issues associated with this PR
N/A

### Describe the problem that is addressed by this PR
The REST API could return an error response with no body. This is
against the Azure design guidelines but it happens in practice.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
N/A

### Are there test cases added in this PR? _(If not, why?)_
Yes

### Provide a list of related PRs _(if any)_
N/A

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [x] Added a changelog (if necessary)