[core] core-util: depend on ts-http-runtime (#33721)
### Packages impacted by this PR

- `@azure/core-util`
- `@typespec/ts-http-runtime`

### Issues associated with this PR

- #32698

### Describe the problem that is addressed by this PR

Hello, friends, I am back again! Now that we have high-level agreement
on this approach, this month I am going through the Core packages and
getting this dedupe train rolling. Starting off with core-util again
since it's simple, but overall it's a similar approach to the last PR,
so check out the description there
(https://github.com/Azure/azure-sdk-for-js/pull/32700). The key
differences from that PR are:
- Not doing the underscores anywhere -- hopefully, the word "internal"
should be enough to make it clear to folks that these are internal
packages. It looks a bit cleaner this way, we can use the `import * as
...` syntax to avoid name conflicts, and I also added a note in the
README that the internal subpath is private and should not be used by
end users.
- I did an analysis of which utilities are actually being used in the
unbranded package and removed everything that wasn't being used. Because
of this, `core-util` still has a handful of utilities; these are the
ones which are not used in the unbranded package.
- Removed tests from `core-util` that were only testing stuff that is
reexported from unbranded. I figure these tests are redundant since we
already have coverage in the other package.