[core] logger: depend on ts-http-runtime (#33752)
### Packages impacted by this PR

- `@azure/core-util`
- `@typespec/ts-http-runtime`

### Issues associated with this PR

- #32698

### Describe the problem that is addressed by this PR

Same effort as https://github.com/Azure/azure-sdk-for-js/pull/33721, but
for the logger this time. The logger is a bit special, as it has some
global state (set/get log level, the root logger, the log level
environment variable name). The approach I took to share the code here
is to wrap the logger implementation in a "context" factory function
which allows Azure and unbranded to have two separate instances of the
state. Azure and unbranded each create an instance of the context and
their exports refer to the context.