[monitor-query][doc] improve `QueryTimeInterval` reference doc (#33909)
The `duration` format is already explained in individual types of the union.
However, the overall doc for the type doesn't include it. This PR adds the info to
the overall ref doc as well so that when ref docs of individual types are not
showing we can still understand the format of the duration.