[core] core-rest-pipeline: depend on ts-http-runtime (#33948)
### Packages impacted by this PR

- `@azure/core-rest-pipeline`
- `@typespec/ts-http-runtime`

### Issues associated with this PR

- #32698

### Describe the problem that is addressed by this PR

Updating core-rest-pipeline to depend on the unbranded package. This one
is a bit more involved than the changes to
[core-util](https://github.com/Azure/azure-sdk-for-js/pull/33721) and
[logger](https://github.com/Azure/azure-sdk-for-js/pull/33752), but
generally follows the same principles.

Specific callouts:
- `AbortSignalLike`. Since we decided that the unbranded package will
only support the native AbortSignal, there is a slight incompatability
between the unbranded and Azure `PipelineRequest`. I worked around this
by casting where necessary, but also added logic to wrap any
`AbortSignalLike` that is not in fact a native `AbortSignal` behind an
`AbortSignal` that reflects the state of the underlying AbortSignalLike.
This wrapping should keep the cast safe at runtime, in case any
unbranded policies or the unbranded HttpClient requires a native
AbortSignal for some reason. This wrapping is done in two places -- in a
new pipeline policy, to ensure that any subsequent policies receive a
native AbortSignal, and in the HttpClient sendRequest function, so that
any requests which aren't passed through a pipeline also get handled.
Would appreciate feedback on this approach.
- `RestError`. To ensure that the unbranded and Azure `RestError`
instances are compatible, the Azure RestError constructor now returns an
unbranded RestError (who knew you could return something in a
constructor? I didn't). This has the effect of making sure we don't have
two versions of RestError lying around at runtime which could cause some
weird behavior.
- User agent policy. I kept the user agent logic duplicated for now
while we figure out what we want to do about #33847. This will probably
involve some refactoring in unbranded to allow customizing the name of
the user agent header for browser/RN.
- Tests. Like for core-util and the logger, I removed tests that were
just testing things that were directly re-exported (after checking that
they passed!). Whenever there's a bit more wrapping, e.g. for the http
clients, I was a bit more conservative and kept most if not all of the
existing tests.