[core] fix stack overflow when passing a Blob for multipart part body in Node (#34160)
### Packages impacted by this PR

- `@typespec/ts-http-runtime`

### Issues associated with this PR

- Fixes bug introduced in #33948 (unreleased)

### Describe the problem that is addressed by this PR

Came across this when investigating with @HarshaNalluru yesterday
evening. I introduced this when taking the `createFile` shim and friends
out of the Unbranded package. The logic I 'fixed' results in infinite
recursion if a `Blob` is passed into `toStream` in Node. Apparently we
didn't have a test for passing a `Blob` body into `multipartPolicy` so
this wasn't caught. I added a test here to fix the gap in coverage.

Ultimately, it seems unlikely that someone would actually pass a Blob as
input here in Node, so the impact of this would probably have been
limited, but it's still a bug and it's an easy fix.