[EngSys] Rename test entry points (#34132)
### Packages impacted by this PR

- All Packages

### Issues associated with this PR


### Describe the problem that is addressed by this PR

This is to propose a rename to our test methods since we do not make
differentiating differences between different types of tests. Instead, I
propose the following:

- `test`: Runs all test suites that can be automatically run
- `test:browser`: Runs the browser tests
- `test:node`: Runs the Node based tests
- `test:node:esm`: Runs the tests on the ESM output

We should be then able to get rid of the `unit-test` and
`integration-test` differences and remove them from our packages. There
may be cases for other tests such as `manual` and `emulator`, but those
can be folded in by named `test:` scripts.

The following packages have integration tests currently that perform
various tasks.

- @azure/cosmos
- @azure/identity
- @azure/notification-hubs
- @azure/service-bus

With these, we propose to merge the tests into a single one if they are
to be executed as part of CI/check-in. Otherwise they will be named
accordingly with `test:node:<type>` or `test:browser:<type>`

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_


### Provide a list of related PRs _(if any)_

- https://github.com/Azure/azure-sdk-for-js/pull/33997

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)