Add support for Weighted RRF +  Hybrid Search Query plan optimization (#34222)
### Packages impacted by this PR
@azure/cosmos

### Issues associated with this PR
#34221


### Describe the problem that is addressed by this PR

1. ***Add support for weighted RRF in hybrid search.***
We allow weights to be negative but the negative sign is used to signal
that we should sort scores in ascending order for the corresponding
component. The final WRRF score is then computed using the absolute
value of the weight.
In this approach, the sign of the weight indicates the interpretation of
the ranking itself rather than directly affecting the calculated score:
WRRF(d) = ∑ |w_i| × 1/(k + r_i'(d))

2. ***Adds support for the optimized query plan***
Adds a QueryFeature that returns optimized query plan, effectively
removing the need for rewriting orderByExpressions in the SDK.
A flag `disableHybridSearchQueryPlanOptimization` is added to disable
returning optimized query plan. This is done to ensure the query works
as expected for older gateways.

### Are there test cases added in this PR? _(If not, why?)_
Yes

### Provide a list of related PRs _(if any)_
* https://github.com/Azure/azure-cosmos-dotnet-v3/pull/5064
* https://github.com/Azure/azure-cosmos-dotnet-v3/pull/5120
* https://github.com/Azure/azure-cosmos-dotnet-v3/pull/5121

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)