[engsys][keyvault] credscan suppression for sample code (#34403)
### Packages impacted by this PR

@azure/keyvault-certificates

### Issues associated with this PR

Fixes #34397

### Describe the problem that is addressed by this PR

Our `importCertificate` sample requires a valid certificate to be run as
part of our sample validation, which provides added confidence in our samples
and package changes.

It is not used by anything, and is simply imported into an ephemeral
resource during tests.

Using string concatenation is the workaround we are using