[EventHubs] fix typecheck tests
in latest version of vitest, the augmenting type declaration of `Object.should`
was commented out, which broke our type checking for `.should` usage in the
test.

This PR adds the declaration in Event Hubs' chai test util and import it once so
that the augmentation is available in global scope.