[EventHubs] fix typecheck tests (#34483)
In latest version of vitest, the augmenting type declaration of `Object.should`
was commented out, which broke our type checking for `.should` usage in
the test.

This PR adds the declaration in Event Hubs' chai test util and imports it once so
that the augmentation is available in global scope.