[Monitor OpenTelemetry Exporter] Should Never Error Log from Statsbeat Init Failure (#34447)
### Packages impacted by this PR
@azure/monitor-opentelemtry-exporter

### Describe the problem that is addressed by this PR
Should not send error logs when failure to init long interval statsbeat
happens. In the past this was causing the SDK to log errors if the daily
cap was hit by our statsbeat LAW.

### Are there test cases added in this PR? _(If not, why?)_
No, just an update to logging level.

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [x] Added a changelog (if necessary)