Updating Communication Common SDK to version supporting Teams Phone Extensibility GA version (#34400)
### Packages impacted by this PR
@azure/communication-common

### Issues associated with this PR


### Describe the problem that is addressed by this PR
ACS Auth TeamsExtension REST API moved to GA.
This PR updates the Communication Common SDK to use the GA version of
the TeamsExtension REST API for token exchange and prepares the SDK for
major release.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_


### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [x] Added a changelog (if necessary)