[core] fix AZURE_LOG_LEVEL environment variable not being properly respected (#34444)
### Packages impacted by this PR

- `@typespec/ts-http-runtime`

### Describe the problem that is addressed by this PR

I noticed that setting `AZURE_LOG_LEVEL` was not updating the log level
properly. This was because when updating the log level to reflect the
environment variable value in `createLoggerContext`, we were calling the
top-level `setLogLevel` instead of the context's setLogLevel function
(so effectively, we were updating the unbranded log level instead of the
Azure one like we should have been). This PR fixes that behavior by
defining `setLogLevel` in `createLoggerContext` instead of defining it
inline at the return site, so that the correct version of `setLogLevel`
is called when we read the environment variable.